Decouple playing a card from ending your turn
Previously, playing a card automatically changed the game state to
give your opponent their turn. However, given that Cribbage requires
players to notice and explicitly claim their points, there must be
an opportunity to do this between playing a card and ending your turn.
Thus, these need to be separate actions.

The current implementation decouples these by adding a CircularState
that tracks the state of the entire circular counting section of a
round of cribbage. It gets stuck right now because there is no action
to end your turn. This will be resolved momentarily.